(from getambassador.io) Emissary 2.4: Overhaul and update the TCPMapping docs
My goal here was to document the changes in 2.4.0, but I quickly
discovered that that I didn't have a stable base to work on.

Everything in my new version was true of older releases as well; except
for the parts about `Listeners` and how it pairs with terminations
`TLSContexts` or `Hosts`.  The pairing-with-TLSContexts restores the 1.x
behavior.  The pairing-with-Hosts more-or-less matches the 2.x behavior (I
say more-or-less because up until now the 2.x behavior was buggy).

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>