(from getambassador.io) emissary: update basic rate-limit tutorial.
This fixes the broken tutorial for a basic rate-limit service. A new
example ratelimit service was created and it behaves exactly the same
as the existing v2 test one did.

It is now stored in its own github repository and has been updated
to support Emissary v3 and Envoy V3 Transport Protocol.

Signed-off-by: Lance Austin <laustin@datawire.io>